pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webscraper"
version = "1.0.0"
description = "A professional, async-capable web scraper with CLI support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Web Scraper Team"}
]
keywords = ["web scraping", "async", "crawler", "data extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
playwright = ["playwright>=1.40.0"]
database = ["sqlalchemy>=2.0.0", "psycopg2-binary>=2.9.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
]
all = ["webscraper[playwright,database,dev]"]

[project.scripts]
webscraper = "webscraper.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/webscraper"
Repository = "https://github.com/yourusername/webscraper"

[tool.setuptools.packages.find]
where = ["."]
include = ["webscraper*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["webscraper"]
